Controller completed
need to correct relationship between tables

+++ src/main/java/com/group13/EmployeeManager/repository/DivisionRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DivisionRepository extends JpaRepository<Division, Integer> {
+public interface DivisionRepository extends JpaRepository<Division, Long> {
     @Query("SELECT d FROM Division d WHERE d.name = :divisionName")
     Division findByDivisionName(String divisionName);
 }
